requirements.txt
gitpython==3.1.43
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
torch==2.2.2
transformers>=4.47.0
accelerate>=0.34.0
scikit-learn==1.4.2
scipy==1.13.0
tqdm==4.66.2
huggingface-hub>=0.26.0

#pip install -U faiss-cpu numpy torch transformers accelerate gitpython sentence-transformers